data-watch/app/requirements.txt
pip-tools
environ-config
pylint==2.4.4
pandas==0.25.3
requests==2.22.0
unidecode==1.1.1
configobj==5.0.6
cycler==0.10.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
psycopg2-binary==2.8.4
py2jdbc==0.0.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12